Dynamic Easiness factor to inform review date

+++ backend/models.py
@@ -1,3 +1,4 @@
+# /Users/nathandiamond/Documents/GitHub/AlgoAscent/backend/models.py
 from typing import Optional
 from sqlmodel import Field, SQLModel
 from datetime import datetime
@@ -17,7 +18,7 @@ class UserAttempt(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     problem_id: int = Field(foreign_key="problem.id")
     attempt_date: Optional[datetime] = Field(default=datetime.now())
-    time_taken_minutes: Optional[int] = None
+    time_taken_minutes: Optional[float] = None
     solved: bool
     notes: Optional[str] = None
     last_attempt_date: Optional[datetime] = Field(default=datetime.now())
